re-delete file
